+++ flask_admin/contrib/geoa/typefmt.py
@@ -17,8 +17,10 @@ def geom_formatter(view, value):
         "data-tile-layer-url": view.tile_layer_url,
         "data-tile-layer-attribution": view.tile_layer_attribution
     })
-    if value.srid is -1:
+
+    if value.srid == -1:
         value.srid = 4326
+
     geojson = view.session.query(view.model).with_entities(func.ST_AsGeoJSON(value)).scalar()
     return Markup('<textarea %s>%s</textarea>' % (params, geojson))